Update Ruff configuration and fix code linting rules
Configure explicit rule selection (E, F, I, UP, C4, RUF) and ignore sets
(E501, RUF001-RUF003, RUF012) in pyproject.toml to incorporate beneficial
quality checks while avoiding CI failures on local date handling and i18n
unicode characters. Fix all linting issues across src and tests.

+++ src/mira/ui/dialogs/financial/goal_simulator.py
@@ -362,7 +362,7 @@ class GoalScenarioDialog(QDialog):
     @staticmethod
     def _target_date_from_years(years: float) -> str:
         """Return target date from years."""
-        days = max(1, int(round(years * 365)))
+        days = max(1, round(years * 365))
         return (date.today() + timedelta(days=days)).isoformat()
 
     def _fmt_currency(self, value: float) -> str:
